Initial commit : Simulation working
The simulation of up to 8192 nodes (CNET does not seem to support more
than ~10k) works.
Scripts to automate simulation execution, topology file generation and
trace parsing are developed.
A diagram of the implemented basic "IP" and Tree protocols is located in
cnet-simulation.drawio.

Next step: Adjust the topology generator to generate LAN segments
instead of the router-connected WAN links. This is to see if the central
router causes avoidable delays. If that is the case, a recommendation to
use bus networks instead of a star topology can be made.